requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
aiosqlite==0.21.0
anyio==4.8.0
APScheduler==3.11.0
attrs==25.3.0
cachetools==5.5.1
certifi==2025.7.14
charset-normalizer==3.4.1
class-registry==2.1.2
filters==1.3.2
frozenlist==1.7.0
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.70.0
gspread==6.0.2
gspread_asyncio==2.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
multidict==6.4.4
numpy==2.2.2
oauthlib==3.2.2
pandas==2.2.3
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.3
pytest-win32consoletitle==0.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-telegram-bot==22.0
pytz==2025.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
sqlite3_api==2.0.4
StrEnum==0.4.15
tenacity==9.0.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
uritemplate==4.2.0
urllib3==2.3.0
watchdog==5.0.3
Win32Security==2.1.0
yarl==1.20.1
